# docs/openapi/ops-api.yaml
openapi: 3.0.3
info:
  title: OpsAI Intelligence Platform API
  description: |
    运维智脑平台后端契约（ops-api）。

    - 响应统一包装为 CommonResult（对齐 EcomAI）
    - 阶段一范围：Health、Auth、Alertmanager Webhook、Alert 列表、Incident CRUD、Timeline
    - 认证：JWT Bearer（`POST /api/v1/auth/login`）；Webhook 端点无需认证
  version: 0.1.0-phase1
  contact:
    name: OpsAI Team

servers:
  - url: http://127.0.0.1:8280
    description: 本地开发（ops-api）

tags:
  - name: Health
    description: 健康检查
  - name: Webhooks
    description: 外部系统回调（Alertmanager）
  - name: Auth
    description: JWT 登录
  - name: Alerts
    description: 告警事件列表（Alert Center）
  - name: Incidents
    description: 故障管理（Dispatch 思路）
  - name: Timeline
    description: Incident 审计时间线

paths:
  /health:
    get:
      tags: [Health]
      summary: 健康检查
      operationId: getHealth
      responses:
        '200':
          description: 服务正常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultHealth'

  /webhooks/alertmanager:
    post:
      tags: [Webhooks]
      summary: 接收 Alertmanager Webhook（v4 JSON）
      description: |
        解析 Prometheus Alertmanager 官方 Webhook 载荷。
        - 写入 `alert_event`
        - Redis fingerprint 去重（TTL 30min）
        - 满足规则时自动创建或归并 `incident`
      operationId: postAlertmanagerWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertmanagerWebhookPayload'
            examples:
              firingCritical:
                summary: 单条 critical 告警（阶段一验收样例）
                value:
                  status: firing
                  receiver: opsai-webhook
                  groupLabels:
                    alertname: HighErrorRate
                  commonLabels:
                    severity: critical
                    service: ecom-api-portal
                  alerts:
                    - status: firing
                      labels:
                        alertname: HighErrorRate
                        service: ecom-api-portal
                        severity: critical
                      annotations:
                        summary: portal 5xx > 5%
                        description: HTTP 5xx error rate exceeds threshold
                      startsAt: '2025-11-11T10:00:00Z'
                      fingerprint: demo-fp-001
      responses:
        '200':
          description: 处理成功（可能新建 incident 或仅归并告警）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultWebhookResult'
        '400':
          description: 载荷格式错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultError'

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 用户登录（JWT）
      operationId: postAuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultLogin'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/alerts:
    get:
      tags: [Alerts]
      summary: 告警事件列表（分页 + 筛选）
      operationId: listAlerts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AlertStatus'
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/Severity'
        - name: source
          in: query
          schema:
            type: string
          example: alertmanager
        - name: service
          in: query
          schema:
            type: string
          example: ecom-api-portal
        - name: fingerprint
          in: query
          description: 精确匹配 fingerprint
          schema:
            type: string
        - name: keyword
          in: query
          description: 标题模糊搜索
          schema:
            type: string
      responses:
        '200':
          description: 分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultAlertPage'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/incidents:
    get:
      tags: [Incidents]
      summary: Incident 列表（分页 + 筛选）
      operationId: listIncidents
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/IncidentStatus'
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/Severity'
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: service
          in: query
          schema:
            type: string
          example: ecom-api-portal
        - name: keyword
          in: query
          description: 标题模糊搜索
          schema:
            type: string
      responses:
        '200':
          description: 分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultIncidentPage'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Incidents]
      summary: 人工创建 Incident
      operationId: createIncident
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultIncident'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/incidents/{incident_id}:
    get:
      tags: [Incidents]
      summary: Incident 详情
      operationId: getIncident
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: 详情（含关联告警摘要）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultIncidentDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Incidents]
      summary: 更新 Incident（状态机动作 / 指派 / 备注）
      description: |
        通过 `action` 驱动状态机，详见 `docs/incident_state_machine.md`。
        非法迁移返回 409。
      operationId: patchIncident
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentPatchRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultIncident'
        '409':
          description: 状态机冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultError'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/incidents/{incident_id}/timeline:
    get:
      tags: [Timeline]
      summary: 获取 Incident 时间线
      operationId: listIncidentTimeline
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 时间线事件列表（按 created_at 降序）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultTimelinePage'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Timeline]
      summary: 添加时间线记录（备注 / 系统外操作留痕）
      operationId: createTimelineEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimelineCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResultTimelineEvent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    IncidentId:
      name: incident_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  responses:
    Unauthorized:
      description: 未认证或 Token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CommonResultError'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CommonResultError'

  schemas:
    # ── CommonResult 包装 ──────────────────────────────────────
    CommonResultHealth:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 操作成功
        data:
          $ref: '#/components/schemas/HealthData'

    HealthData:
      type: object
      required: [status, service, version]
      properties:
        status:
          type: string
          enum: [ok, degraded]
        service:
          type: string
          example: ops-api
        version:
          type: string
          example: 0.1.0-phase1

    CommonResultError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 业务错误码（非 HTTP 状态码）
          example: 400
        message:
          type: string
        data:
          type: object
          additionalProperties: true
          nullable: true

    # ── 枚举 ──────────────────────────────────────────────────
    Severity:
      type: string
      enum: [critical, high, medium, low]

    IncidentStatus:
      type: string
      enum:
        - open
        - acknowledged
        - investigating
        - mitigated
        - resolved
        - closed

    IncidentAction:
      type: string
      enum:
        - acknowledge
        - start_investigation
        - mitigate
        - resolve
        - close
        - reopen_investigation
        - assign
        - update_severity
        - add_note

    TimelineEventType:
      type: string
      enum:
        - status_change
        - assignment
        - severity_change
        - note
        - alert_merged
        - system

    ActorType:
      type: string
      enum: [user, system]

    AlertStatus:
      type: string
      enum: [firing, resolved]

    # ── Alertmanager Webhook ──────────────────────────────────
    AlertmanagerWebhookPayload:
      type: object
      required: [status, alerts]
      properties:
        status:
          type: string
          enum: [firing, resolved]
        receiver:
          type: string
        groupLabels:
          type: object
          additionalProperties:
            type: string
        commonLabels:
          type: object
          additionalProperties:
            type: string
        commonAnnotations:
          type: object
          additionalProperties:
            type: string
        externalURL:
          type: string
          format: uri
        alerts:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AlertmanagerAlert'

    AlertmanagerAlert:
      type: object
      required: [status, labels, fingerprint]
      properties:
        status:
          $ref: '#/components/schemas/AlertStatus'
        labels:
          type: object
          additionalProperties:
            type: string
        annotations:
          type: object
          additionalProperties:
            type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
          nullable: true
        generatorURL:
          type: string
          format: uri
        fingerprint:
          type: string
          description: 告警唯一指纹，用于去重与归并

    CommonResultWebhookResult:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
        data:
          $ref: '#/components/schemas/WebhookResult'

    WebhookResult:
      type: object
      required: [processed_count, alert_event_ids]
      properties:
        processed_count:
          type: integer
        alert_event_ids:
          type: array
          items:
            type: integer
            format: int64
        incident_id:
          type: integer
          format: int64
          nullable: true
          description: 新建或归并到的 Incident；纯去重时可为 null
        incident_created:
          type: boolean
          description: 本次是否新建了 Incident

    # ── Incident ────────────────────────────────────────────
    IncidentSummary:
      type: object
      required:
        - id
        - incident_no
        - title
        - status
        - severity
        - service
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        incident_no:
          type: string
          example: INC-20251111-0001
        title:
          type: string
        status:
          $ref: '#/components/schemas/IncidentStatus'
        severity:
          $ref: '#/components/schemas/Severity'
        service:
          type: string
          nullable: true
          example: ecom-api-portal
        owner_id:
          type: integer
          format: int64
          nullable: true
        owner_name:
          type: string
          nullable: true
          description: 负责人显示名（列表展示用）
        root_cause_preview:
          type: string
          nullable: true
          description: 根因摘要（阶段三 RCA 写入；阶段一可为 null）
        alert_count:
          type: integer
          minimum: 0
          description: 关联告警数量
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true

    Incident:
      allOf:
        - $ref: '#/components/schemas/IncidentSummary'
        - type: object
          properties:
            description:
              type: string
              nullable: true
            primary_fingerprint:
              type: string
              nullable: true
              description: 首条告警 fingerprint

    IncidentDetail:
      allOf:
        - $ref: '#/components/schemas/Incident'
        - type: object
          properties:
            related_alerts:
              type: array
              items:
                $ref: '#/components/schemas/AlertEventSummary'
            recent_timeline:
              type: array
              description: 最近 5 条时间线（详情页预览）
              items:
                $ref: '#/components/schemas/TimelineEvent'

    AlertEventSummary:
      type: object
      required: [id, fingerprint, severity, title, status, created_at]
      properties:
        id:
          type: integer
          format: int64
        fingerprint:
          type: string
        severity:
          $ref: '#/components/schemas/Severity'
        title:
          type: string
        status:
          $ref: '#/components/schemas/AlertStatus'
        source:
          type: string
          example: alertmanager
        created_at:
          type: string
          format: date-time

    IncidentCreateRequest:
      type: object
      required: [title, severity]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 256
        description:
          type: string
          maxLength: 4000
        severity:
          $ref: '#/components/schemas/Severity'
        service:
          type: string
          maxLength: 128
        owner_id:
          type: integer
          format: int64

    IncidentPatchRequest:
      type: object
      properties:
        action:
          $ref: '#/components/schemas/IncidentAction'
        owner_id:
          type: integer
          format: int64
          description: 与 action=assign 配合，或直接指派
        severity:
          $ref: '#/components/schemas/Severity'
          description: 与 action=update_severity 配合
        note:
          type: string
          maxLength: 2000
          description: 与 action=add_note 配合，或附在状态变更上的备注
        title:
          type: string
          maxLength: 256
        description:
          type: string
          maxLength: 4000

    CommonResultIncident:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
        data:
          $ref: '#/components/schemas/Incident'

    CommonResultIncidentDetail:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/IncidentDetail'

    PageMeta:
      type: object
      required: [page, page_size, total, total_pages]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
          format: int64
        total_pages:
          type: integer

    IncidentPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IncidentSummary'
        meta:
          $ref: '#/components/schemas/PageMeta'

    CommonResultIncidentPage:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/IncidentPage'

    # ── Timeline ──────────────────────────────────────────────
    TimelineEvent:
      type: object
      required: [id, incident_id, event_type, content, actor_type, created_at]
      properties:
        id:
          type: integer
          format: int64
        incident_id:
          type: integer
          format: int64
        event_type:
          $ref: '#/components/schemas/TimelineEventType'
        content:
          type: string
        actor_type:
          $ref: '#/components/schemas/ActorType'
        actor_id:
          type: integer
          format: int64
          nullable: true
        actor_name:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
          description: 如 from_status / to_status / alert_id
        created_at:
          type: string
          format: date-time

    TimelineCreateRequest:
      type: object
      required: [content]
      properties:
        event_type:
          $ref: '#/components/schemas/TimelineEventType'
          default: note
        content:
          type: string
          minLength: 1
          maxLength: 2000
        metadata:
          type: object
          additionalProperties: true

    TimelinePage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEvent'
        meta:
          $ref: '#/components/schemas/PageMeta'

    CommonResultTimelinePage:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/TimelinePage'

    CommonResultTimelineEvent:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/TimelineEvent'

    # ── Auth ──────────────────────────────────────────────────
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: OpsAI@2025

    LoginData:
      type: object
      required: [access_token, token_type, user_id, username, display_name, role]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        user_id:
          type: integer
          format: int64
        username:
          type: string
        display_name:
          type: string
        role:
          type: string
          enum: [admin, operator]

    CommonResultLogin:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
        data:
          $ref: '#/components/schemas/LoginData'

    # ── Alert 列表 ────────────────────────────────────────────
    AlertPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AlertEventSummary'
        meta:
          $ref: '#/components/schemas/PageMeta'

    CommonResultAlertPage:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/AlertPage'